Files resulted from the semantic classification of 4 models
I have used four models to determine which one is the best for labeling data semantically. 
labeled_insurance_companies.csv is the result of the all-MiniLM-L6 model
labeled_insurance_companies_2.csv is the result of the all-mpnet-base-v2 model
labeled_insurance_companies_3.csv is the result of the all-roberta-large-v1 model
labeled_insurance_companies_4.csv is the result of the multi-qa-mpnet-base-dot-v1 model

The multi-qa-mpnet-base-dot-v1 model has labelled the data semantically the best compared to the other models.